Move Python version specific decorator code to packages.
Add switch in setup.py to only install either the Python2 or Python3 version to avoid bytecompile errors during packaging.

Addresses #759
