+++ networkx/external/decorator/decorator2/__init__.py
@@ -1,0 +1,1 @@
+#
