+++ networkx/external/decorator/decorator2/_decorator2.py
