+++ networkx/external/decorator/decorator3/_decorator3.py
